Fix -b default value overriding consumer config
The default value of the command line argument `-b` (`--bootstrap-brokers`) overrides the value given in the consumer config (`consumer.properties`) file as the code can't tell if the value is the default or it was provided. This problem manifests when `-b` is not passed as argument and `consumer.properties` contains `bootstrap.server` with a value other than `localhost`. If not overrided by argument, the value in the `consumer.properties` file should be used.

The fix involves dropping the default value in the arguments definition to allow the code to later overwrite the config if the argument was really present. The default value for the bootstrap brokers is set in the initialization of the `consumer_config` dict maintaining the behavior stated in the docs.